triton/model-api/requirements.txt
fastapi==0.103.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==2.1.1
pyjwt==2.8.0
minio==7.1.15
kubernetes==26.1.0
aiohttp==3.8.5
jinja2==3.1.2
python-dotenv==1.0.0
pytest==7.4.0
httpx==0.24.1
